LayoutTests/fast/css/style-scoped/style-scoped-attach-expected.txt

Test attaching a new <style scoped> element

Text
Text

Text
Text

Text
Text

--- ORIGINAL ---
S1: rgb(0, 0, 0)
P1: rgb(0, 0, 0)
S2: rgb(0, 0, 0)
P2: rgb(0, 0, 0)
S3: rgb(0, 0, 0)
P3: rgb(0, 0, 0)
--- AFTER ATTACHING <STYLE SCOPED> ---
S1: rgb(0, 0, 0)
P1: rgb(0, 0, 0)
S2: rgb(255, 0, 0)
P2: rgb(0, 128, 0)
S3: rgb(0, 0, 0)
P3: rgb(0, 0, 0)
--- FINISHED ---
